remove AWS region from helm chart (#5567)
# Description
This PR removes AWS region from helm charts and sets the right region
for AWS requests made by cloudcontrol and cloud formation clients based
on the region indicated by the request URL.

## Issue reference
Fixes: #5493

## Checklist

Please make sure you've completed the relevant tasks for this PR, out of
the following list:

* [ X] Code compiles correctly
* [ X] Adds necessary unit tests for change
* [ X] Adds necessary E2E tests for change
* [ X] Unit tests passing
* [ NA] Extended the documentation / Created issue for it

+++ pkg/ucp/frontend/controller/awsproxy/deleteawsresource.go
@@ -43,11 +43,16 @@ func NewDeleteAWSResource(opts ctrl.Options) (armrpc_controller.Controller, erro
 
 func (p *DeleteAWSResource) Run(ctx context.Context, w http.ResponseWriter, req *http.Request) (armrpc_rest.Response, error) {
 	serviceCtx := servicecontext.AWSRequestContextFromContext(ctx)
+	region, errResponse := readRegionFromRequest(req.URL.Path, p.basePath)
+	if errResponse != nil {
+		return errResponse, nil
+	}
 
+	cloudControlOpts := []func(*cloudcontrol.Options){CloudControlRegionOption(region)}
 	response, err := p.awsOptions.AWSCloudControlClient.DeleteResource(ctx, &cloudcontrol.DeleteResourceInput{
 		TypeName:   to.Ptr(serviceCtx.ResourceTypeInAWSFormat()),
 		Identifier: aws.String(serviceCtx.ResourceID.Name()),
-	})
+	}, cloudControlOpts...)
 	if err != nil {
 		if awsclient.IsAWSResourceNotFoundError(err) {
 			return armrpc_rest.NewNoContentResponse(), nil
